requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.0.1
click==8.1.7
fastapi-slim==0.111.0
h11==0.14.0
idna==3.7
passlib==1.7.4
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.1
